+++ Fun.Blazor/CHANGELOG.md
@@ -13,6 +13,7 @@
 - Support to get value for other type in AdaptiveForm
 - Do not increase index for emptyAttr or emptyNode
 - Support CE empty body for fsharp 9
+- Expose component instance from IComponentHook
 
 ## [4.0.12] - 2024-06-05